fix compiling

+++ contracts/test/OracleEchidnaTest.sol
@@ -62,7 +62,7 @@ contract OracleEchidnaTest {
         (int56[] memory tickCumulatives, uint160[] memory secondsPerLiquidityCumulativeX128s) =
             oracle.observe(secondsAgos);
         int56 timeWeightedTick = (tickCumulatives[1] - tickCumulatives[0]) / timeElapsed;
-        uint160 timeWeightedHarmonicMeanLiquidity =
+        uint256 timeWeightedHarmonicMeanLiquidity =
             ((uint256(timeElapsed) << 160) /
                 (secondsPerLiquidityCumulativeX128s[1] - secondsPerLiquidityCumulativeX128s[0])) >> 32;
         assert(timeWeightedHarmonicMeanLiquidity <= type(uint128).max);
@@ -128,7 +128,7 @@ contract OracleEchidnaTest {
         // the time weighted averages fit in their respective accumulated types
         assert(timeWeightedTick <= type(int24).max && timeWeightedTick >= type(int24).min);
 
-        uint160 timeWeightedHarmonicMeanLiquidity =
+        uint256 timeWeightedHarmonicMeanLiquidity =
             ((uint256(secondsAgo) << 160) /
                 (secondsPerLiquidityCumulativeX128s[1] - secondsPerLiquidityCumulativeX128s[0])) >> 32;
